Change another bunch of shit
- Add network permission to manifest
- Disable strict mode's no network on main thread policy
- Change some floats to doubles
- Change UI to a dark theme feel
- Fix periodic table instancing and text listeners

+++ app/src/main/java/com/example/aps/PeriodicElement.java
@@ -6,8 +6,8 @@ public class PeriodicElement
     public String symbol;
     public String name;
     public String mass;
-    public float meltingPoint;
-    public float boilingPoint;
+    public double meltingPoint;
+    public double boilingPoint;
     public boolean fetched = false;
 
     public PeriodicElement() { }
